+++ tests/subscriptions/test_conf.py
@@ -4,7 +4,7 @@ from decimal import Decimal
 from django.conf import settings
 from django.test import override_settings
 
-from subscriptions.conf import compile_settings, Currency, CURRENCY
+from subscriptions import conf
 
 
 @override_settings(
@@ -12,17 +12,20 @@ from subscriptions.conf import compile_settings, Currency, CURRENCY
     DFS_CURRENCY_LOCALE=2,
     DFS_BASE_TEMPLATE='3',
     DFS_SUBSCRIBE_VIEW='a.b',
+    DFS_MANAGER_CLASS='a.b',
 )
 def test_all_settings_populate_from_settings_properly():
     """Tests that Django settings all proper populate SETTINGS."""
-    subscription_settings = compile_settings()
+    subscription_settings = conf.compile_settings()
 
     assert len(subscription_settings) == 5
     assert subscription_settings['enable_admin'] == 1
     assert subscription_settings['currency_locale'] == '2'
     assert subscription_settings['base_template'] == '3'
-    assert subscription_settings['subscribe_view_module'] == 'a'
-    assert subscription_settings['subscribe_view_class'] == 'b'
+    assert subscription_settings['subscribe_view']['module'] == 'a'
+    assert subscription_settings['subscribe_view']['class'] == 'b'
+    assert subscription_settings['management_manager']['module'] == 'a'
+    assert subscription_settings['management_manager']['class'] == 'b'
 
 @override_settings()
 def test_settings_defaults():
@@ -32,25 +35,36 @@ def test_settings_defaults():
     del settings.DFS_CURRENCY_LOCALE
     del settings.DFS_BASE_TEMPLATE
     del settings.DFS_SUBSCRIBE_VIEW
+    del settings.DFS_MANAGER_CLASS
 
-    subscription_settings = compile_settings()
+    subscription_settings = conf.compile_settings()
 
     assert len(subscription_settings) == 5
     assert subscription_settings['enable_admin'] is False
     assert subscription_settings['currency_locale'] == 'en_us'
     assert subscription_settings['base_template'] == 'subscriptions/base.html'
-    assert subscription_settings['subscribe_view_module'] == 'subscriptions.views'
-    assert subscription_settings['subscribe_view_class'] == 'SubscribeView'
+    assert subscription_settings['subscribe_view']['module'] == (
+        'subscriptions.views'
+    )
+    assert subscription_settings['subscribe_view']['class'] == (
+        'SubscribeView'
+    )
+    assert subscription_settings['management_manager']['module'] == (
+        'subscriptions.management.commands._manager'
+    )
+    assert subscription_settings['management_manager']['class'] == (
+        'Manager'
+    )
 
 def test_format_currency_en_us():
     """Tests that format currency works properly with the default."""
-    currency_value = CURRENCY['en_us'].format_currency('1234567.987')
+    currency_value = conf.CURRENCY['en_us'].format_currency('1234567.987')
 
     assert currency_value == '$1,234,567.99'
 
 def test_format_currency_rounding_string():
     """Tests that rounding works as expected with string values."""
-    test_currency = Currency()
+    test_currency = conf.Currency()
 
     assert test_currency.format_currency('0.506') == '0.51'
     assert test_currency.format_currency('0.505') == '0.51'
@@ -62,7 +76,7 @@ def test_format_currency_rounding_string():
 
 def test_format_currency_rounding_decimal():
     """Tests that rounding works as expected with decimal values."""
-    test_currency = Currency()
+    test_currency = conf.Currency()
 
     assert test_currency.format_currency(Decimal('0.506')) == '0.51'
     assert test_currency.format_currency(Decimal('0.505')) == '0.51'
@@ -74,7 +88,7 @@ def test_format_currency_rounding_decimal():
 
 def test_currency_format_grouping_by_1():
     """Tests that grouping works properly for groups of 1."""
-    test_currency = Currency(mon_grouping=1, mon_thousands_sep=',')
+    test_currency = conf.Currency(mon_grouping=1, mon_thousands_sep=',')
 
     assert test_currency.format_currency('1.00') == '1.00'
     assert test_currency.format_currency('10.00') == '1,0.00'
@@ -83,7 +97,7 @@ def test_currency_format_grouping_by_1():
 
 def test_currency_format_grouping_by_3():
     """Tests that grouping works properly for groups of 3."""
-    test_currency = Currency(mon_grouping=3, mon_thousands_sep=',')
+    test_currency = conf.Currency(mon_grouping=3, mon_thousands_sep=',')
 
     assert test_currency.format_currency('1.00') == '1.00'
     assert test_currency.format_currency('10.00') == '10.00'
@@ -95,37 +109,37 @@ def test_currency_format_grouping_by_3():
 
 def test_currency_format_grouping_separator():
     """Tests that the grouping separator is properly applied."""
-    test_currency = Currency(mon_grouping=3, mon_thousands_sep='*')
+    test_currency = conf.Currency(mon_grouping=3, mon_thousands_sep='*')
 
     assert test_currency.format_currency('1000000.00') == '1*000*000.00'
 
 def test_format_currency_symbol_precedes_positive():
     """Tests currency symbol is applied preceding positive value."""
-    test_currency = Currency(currency_symbol='$', p_cs_precedes=True)
+    test_currency = conf.Currency(currency_symbol='$', p_cs_precedes=True)
 
     assert test_currency.format_currency('1.00') == '$1.00'
 
 def test_format_currency_symbol_precedes_negative():
     """Tests currency symbol is applied preceding negative value."""
-    test_currency = Currency(currency_symbol='$', n_cs_precedes=True)
+    test_currency = conf.Currency(currency_symbol='$', n_cs_precedes=True)
 
     assert test_currency.format_currency('-1.00') == '$1.00'
 
 def test_format_currency_symbol_follows_positive():
     """Tests currency symbol is applied following positive value."""
-    test_currency = Currency(currency_symbol='$', p_cs_precedes=False)
+    test_currency = conf.Currency(currency_symbol='$', p_cs_precedes=False)
 
     assert test_currency.format_currency('1.00') == '1.00$'
 
 def test_format_currency_symbol_follows_negative():
     """Tests currency symbol is applied following negative value."""
-    test_currency = Currency(currency_symbol='$', n_cs_precedes=False)
+    test_currency = conf.Currency(currency_symbol='$', n_cs_precedes=False)
 
     assert test_currency.format_currency('-1.00') == '1.00$'
 
 def test_format_currency_symbol_precedes_with_space_positive():
     """Tests currency symbol is applied preceding positive value with space."""
-    test_currency = Currency(
+    test_currency = conf.Currency(
         currency_symbol='$', p_cs_precedes=True, p_sep_by_space=True
     )
 
@@ -133,7 +147,7 @@ def test_format_currency_symbol_precedes_with_space_positive():
 
 def test_format_currency_symbol_precedes_with_space_negative():
     """Tests currency symbol is applied preceding negative value with space."""
-    test_currency = Currency(
+    test_currency = conf.Currency(
         currency_symbol='$', n_cs_precedes=True, n_sep_by_space=True
     )
 
@@ -141,7 +155,7 @@ def test_format_currency_symbol_precedes_with_space_negative():
 
 def test_format_currency_symbol_follows_with_space_positive():
     """Tests currency symbol is applied following positive value with space."""
-    test_currency = Currency(
+    test_currency = conf.Currency(
         currency_symbol='$', p_cs_precedes=False, p_sep_by_space=True
     )
 
@@ -149,7 +163,7 @@ def test_format_currency_symbol_follows_with_space_positive():
 
 def test_format_currency_symbol_follows_with_space_negative():
     """Tests currency symbol is applied following negative value with space."""
-    test_currency = Currency(
+    test_currency = conf.Currency(
         currency_symbol='$', n_cs_precedes=False, n_sep_by_space=True
     )
 
@@ -157,7 +171,7 @@ def test_format_currency_symbol_follows_with_space_negative():
 
 def test_format_currency_symbol_international():
     """Tests that international currency symbol works properly."""
-    test_currency = Currency(int_curr_symbol='USD')
+    test_currency = conf.Currency(int_curr_symbol='USD')
 
     currency_string = test_currency.format_currency('1.00', international=True)
 
@@ -165,7 +179,7 @@ def test_format_currency_symbol_international():
 
 def test_format_currency_sign_position_0_positive_preceding_cs():
     """Tests sign position 0, positive value, and preceding symbol."""
-    test_currency = Currency(
+    test_currency = conf.Currency(
         currency_symbol='$', positive_sign='+', p_sign_posn=0
     )
 
@@ -173,7 +187,7 @@ def test_format_currency_sign_position_0_positive_preceding_cs():
 
 def test_format_currency_sign_position_0_negative_preceding_cs():
     """Tests sign position 0, negative value, and preceding symbol."""
-    test_currency = Currency(
+    test_currency = conf.Currency(
         currency_symbol='$', negative_sign='-', n_sign_posn=0
     )
 
@@ -181,7 +195,7 @@ def test_format_currency_sign_position_0_negative_preceding_cs():
 
 def test_format_currency_sign_position_0_positive_following_cs():
     """Tests sign position 0, positive value, and following symbol."""
-    test_currency = Currency(
+    test_currency = conf.Currency(
         currency_symbol='$', p_cs_precedes=False,
         positive_sign='+', p_sign_posn=0
     )
@@ -190,7 +204,7 @@ def test_format_currency_sign_position_0_positive_following_cs():
 
 def test_format_currency_sign_position_0_negative_following_cs():
     """Tests sign position 0, negative value, and following symbol."""
-    test_currency = Currency(
+    test_currency = conf.Currency(
         currency_symbol='$', n_cs_precedes=False,
         negative_sign='-', n_sign_posn=0
     )
@@ -199,7 +213,7 @@ def test_format_currency_sign_position_0_negative_following_cs():
 
 def test_format_currency_sign_position_1_positive_preceding_cs():
     """Tests sign position 1, positive value, and preceding symbol."""
-    test_currency = Currency(
+    test_currency = conf.Currency(
         currency_symbol='$', positive_sign='+', p_sign_posn=1
     )
 
@@ -207,7 +221,7 @@ def test_format_currency_sign_position_1_positive_preceding_cs():
 
 def test_format_currency_sign_position_1_negative_preceding_cs():
     """Tests sign position 1, negative value, and preceding symbol."""
-    test_currency = Currency(
+    test_currency = conf.Currency(
         currency_symbol='$', negative_sign='-', n_sign_posn=1
     )
 
@@ -215,7 +229,7 @@ def test_format_currency_sign_position_1_negative_preceding_cs():
 
 def test_format_currency_sign_position_1_positive_following_cs():
     """Tests sign position 1, positive value, and following symbol."""
-    test_currency = Currency(
+    test_currency = conf.Currency(
         currency_symbol='$', p_cs_precedes=False,
         positive_sign='+', p_sign_posn=1
     )
@@ -224,7 +238,7 @@ def test_format_currency_sign_position_1_positive_following_cs():
 
 def test_format_currency_sign_position_1_negative_following_cs():
     """Tests sign position 1, negative value, and following symbol."""
-    test_currency = Currency(
+    test_currency = conf.Currency(
         currency_symbol='$', n_cs_precedes=False,
         negative_sign='-', n_sign_posn=1
     )
@@ -233,7 +247,7 @@ def test_format_currency_sign_position_1_negative_following_cs():
 
 def test_format_currency_sign_position_2_positive_preceding_cs():
     """Tests sign position 2, positive value, and preceding symbol."""
-    test_currency = Currency(
+    test_currency = conf.Currency(
         currency_symbol='$', positive_sign='+', p_sign_posn=2
     )
 
@@ -241,7 +255,7 @@ def test_format_currency_sign_position_2_positive_preceding_cs():
 
 def test_format_currency_sign_position_2_negative_preceding_cs():
     """Tests sign position 2, negative value, and preceding symbol."""
-    test_currency = Currency(
+    test_currency = conf.Currency(
         currency_symbol='$', negative_sign='-', n_sign_posn=2
     )
 
@@ -249,7 +263,7 @@ def test_format_currency_sign_position_2_negative_preceding_cs():
 
 def test_format_currency_sign_position_2_positive_following_cs():
     """Tests sign position 2, positive value, and following symbol."""
-    test_currency = Currency(
+    test_currency = conf.Currency(
         currency_symbol='$', p_cs_precedes=False,
         positive_sign='+', p_sign_posn=2
     )
@@ -258,7 +272,7 @@ def test_format_currency_sign_position_2_positive_following_cs():
 
 def test_format_currency_sign_position_2_negative_following_cs():
     """Tests sign position 2, negative value, and following symbol."""
-    test_currency = Currency(
+    test_currency = conf.Currency(
         currency_symbol='$', n_cs_precedes=False,
         negative_sign='-', n_sign_posn=2
     )
@@ -267,7 +281,7 @@ def test_format_currency_sign_position_2_negative_following_cs():
 
 def test_format_currency_sign_position_3_positive_preceding_cs():
     """Tests sign position 3, positive value, and preceding symbol."""
-    test_currency = Currency(
+    test_currency = conf.Currency(
         currency_symbol='$', positive_sign='+', p_sign_posn=3
     )
 
@@ -275,7 +289,7 @@ def test_format_currency_sign_position_3_positive_preceding_cs():
 
 def test_format_currency_sign_position_3_negative_preceding_cs():
     """Tests sign position 3, negative value, and preceding symbol."""
-    test_currency = Currency(
+    test_currency = conf.Currency(
         currency_symbol='$', negative_sign='-', n_sign_posn=3
     )
 
@@ -283,7 +297,7 @@ def test_format_currency_sign_position_3_negative_preceding_cs():
 
 def test_format_currency_sign_position_3_positive_following_cs():
     """Tests sign position 3, positive value, and following symbol."""
-    test_currency = Currency(
+    test_currency = conf.Currency(
         currency_symbol='$', p_cs_precedes=False,
         positive_sign='+', p_sign_posn=3
     )
@@ -292,7 +306,7 @@ def test_format_currency_sign_position_3_positive_following_cs():
 
 def test_format_currency_sign_position_3_negative_following_cs():
     """Tests sign position 3, negative value, and following symbol."""
-    test_currency = Currency(
+    test_currency = conf.Currency(
         currency_symbol='$', n_cs_precedes=False,
         negative_sign='-', n_sign_posn=3
     )
@@ -301,7 +315,7 @@ def test_format_currency_sign_position_3_negative_following_cs():
 
 def test_format_currency_sign_position_4_positive_preceding_cs():
     """Tests sign position 4, positive value, and preceding symbol."""
-    test_currency = Currency(
+    test_currency = conf.Currency(
         currency_symbol='$', positive_sign='+', p_sign_posn=4
     )
 
@@ -309,7 +323,7 @@ def test_format_currency_sign_position_4_positive_preceding_cs():
 
 def test_format_currency_sign_position_4_negative_preceding_cs():
     """Tests sign position 4, positive value, and preceding symbol."""
-    test_currency = Currency(
+    test_currency = conf.Currency(
         currency_symbol='$', negative_sign='-', n_sign_posn=4
     )
 
@@ -317,7 +331,7 @@ def test_format_currency_sign_position_4_negative_preceding_cs():
 
 def test_format_currency_sign_position_4_positive_following_cs():
     """Tests sign position 4, positive value, and following symbol."""
-    test_currency = Currency(
+    test_currency = conf.Currency(
         currency_symbol='$', p_cs_precedes=False,
         positive_sign='+', p_sign_posn=4
     )
@@ -326,7 +340,7 @@ def test_format_currency_sign_position_4_positive_following_cs():
 
 def test_format_currency_sign_position_4_negative_following_cs():
     """Tests sign position 4, negative value, and following symbol."""
-    test_currency = Currency(
+    test_currency = conf.Currency(
         currency_symbol='$', n_cs_precedes=False,
         negative_sign='-', n_sign_posn=4
     )
@@ -335,7 +349,7 @@ def test_format_currency_sign_position_4_negative_following_cs():
 
 def test_format_currency_sign_position_other_positive_preceding_cs():
     """Tests 'other' sign position, positive value, and preceding symbol."""
-    test_currency = Currency(
+    test_currency = conf.Currency(
         currency_symbol='$', positive_sign='+', p_sign_posn=None
     )
 
@@ -343,7 +357,7 @@ def test_format_currency_sign_position_other_positive_preceding_cs():
 
 def test_format_currency_sign_position_other_negative_preceding_cs():
     """Tests 'other' sign position, negative value, and preceding symbol."""
-    test_currency = Currency(
+    test_currency = conf.Currency(
         currency_symbol='$', negative_sign='-', n_sign_posn=None
     )
 
@@ -351,7 +365,7 @@ def test_format_currency_sign_position_other_negative_preceding_cs():
 
 def test_format_currency_sign_position_other_positive_following_cs():
     """Tests 'other' sign position, positive value, and following symbol."""
-    test_currency = Currency(
+    test_currency = conf.Currency(
         currency_symbol='$', p_cs_precedes=False,
         positive_sign='+', p_sign_posn=None
     )
@@ -360,9 +374,25 @@ def test_format_currency_sign_position_other_positive_following_cs():
 
 def test_format_currency_sign_position_other_negative_following_cs():
     """Tests 'other' sign position, negative value, and following symbol."""
-    test_currency = Currency(
+    test_currency = conf.Currency(
         currency_symbol='$', n_cs_precedes=False,
         negative_sign='-', n_sign_posn=None
     )
 
     assert test_currency.format_currency('-1.00') == '-1.00$'
+
+def test_string_to_module_and_class_one_period():
+    """Tests handling of string with single period."""
+    string = 'a.b'
+    components = conf.string_to_module_and_class(string)
+
+    assert components['module'] == 'a'
+    assert components['class'] == 'b'
+
+def test_string_to_module_and_class_2_period():
+    """Tests handling of string with more than one period."""
+    string = 'a.b.c'
+    components = conf.string_to_module_and_class(string)
+
+    assert components['module'] == 'a.b'
+    assert components['class'] == 'c'